source/requirements.txt
# ============= 桌面版依赖 =============
# 桌面GUI框架
customtkinter==5.2.1
Pillow==10.2.0
CTkMessagebox==2.1

# ============= 核心功能依赖 =============
# 加密库
cryptography==43.0.0

# 文档处理
python-docx==1.1.2
pdfplumber==0.10.4

# ============= Web版本依赖 =============
# Web框架
streamlit==1.36.0

# ============= 开发和测试依赖 =============
pytest==7.4.3
pytest-cov==4.1.0
